Have async cache implement the interface, remove sync cache
