--- core/server/master/src/test/java/alluxio/master/file/meta/UfsAbsentPathSyncCacheTest.java
@@ -1,207 +1,0 @@
-/*
- * The Alluxio Open Foundation licenses this work under the Apache License, version 2.0
- * (the "License"). You may not use this work except in compliance with the License, which is
- * available at www.apache.org/licenses/LICENSE-2.0
- *
- * This software is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied, as more fully set forth in the License.
- *
- * See the NOTICE file distributed with this work for information regarding copyright ownership.
- */
-
-package alluxio.master.file.meta;
-
-import alluxio.AlluxioURI;
-import alluxio.master.file.options.MountOptions;
-import alluxio.underfs.MasterUfsManager;
-import alluxio.underfs.UfsManager;
-import alluxio.util.IdUtils;
-
-import com.google.common.io.Files;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-
-import java.io.File;
-import java.util.Collections;
-
-/**
- * Unit tests for {@link UfsAbsentPathSyncCache}.
- */
-public class UfsAbsentPathSyncCacheTest {
-  private UfsAbsentPathSyncCache mUfsAbsentPathSyncCache;
-  private MountTable mMountTable;
-
-  private long mMountId;
-  private UfsManager mUfsManager;
-  private String mLocalUfsPath;
-
-  /**
-   * Sets up a new {@link UfsAbsentPathSyncCache} before a test runs.
-   */
-  @Before
-  public void before() throws Exception {
-    mLocalUfsPath = Files.createTempDir().getAbsolutePath();
-    mUfsManager = new MasterUfsManager();
-    mMountTable = new MountTable(mUfsManager);
-    mUfsAbsentPathSyncCache = new UfsAbsentPathSyncCache(mMountTable);
-
-    mMountId = IdUtils.getRandomNonNegativeLong();
-    mUfsManager.addMount(mMountId, mLocalUfsPath, Collections.<String, String>emptyMap());
-    mMountTable.add(new AlluxioURI("/mnt"), new AlluxioURI(mLocalUfsPath), mMountId,
-        MountOptions.defaults());
-  }
-
-  @Test
-  public void isAbsentRoot() throws Exception {
-    // /mnt/a will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI("/mnt/a/b"));
-    checkAbsentPaths(new AlluxioURI("/mnt/a"));
-
-    // /mnt/a will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI("/mnt/a/b/c"));
-    checkAbsentPaths(new AlluxioURI("/mnt/a"));
-
-    // /mnt/1 will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI("/mnt/1/2"));
-    checkAbsentPaths(new AlluxioURI("/mnt/1"));
-
-    // /mnt/1 will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI("/mnt/1/3"));
-    checkAbsentPaths(new AlluxioURI("/mnt/1"));
-  }
-
-  @Test
-  public void isAbsentDirectory() throws Exception {
-    String ufsBase = "/a/b";
-    String alluxioBase = "/mnt" + ufsBase;
-    // Create ufs directories
-    Assert.assertTrue((new File(mLocalUfsPath + ufsBase)).mkdirs());
-
-    // 'base + /c' will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI(alluxioBase + "/c/d"));
-    checkAbsentPaths(new AlluxioURI(alluxioBase + "/c"));
-
-    // 'base + /c' will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI(alluxioBase + "/c/d/e"));
-    checkAbsentPaths(new AlluxioURI(alluxioBase + "/c"));
-
-    // '/a/1' will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI("/mnt/a/1/2"));
-    checkAbsentPaths(new AlluxioURI("/mnt/a/1"));
-
-    // '/1' will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI("/mnt/1/2"));
-    checkAbsentPaths(new AlluxioURI("/mnt/1"));
-  }
-
-  @Test
-  public void isAbsentAddUfsDirectory() throws Exception {
-    String ufsBase = "/a/b";
-    String alluxioBase = "/mnt" + ufsBase;
-    // Create ufs directories
-    Assert.assertTrue((new File(mLocalUfsPath + ufsBase)).mkdirs());
-
-    // 'base + /c' will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI(alluxioBase + "/c/d/e"));
-    checkAbsentPaths(new AlluxioURI(alluxioBase + "/c"));
-
-    // Create a sub-directory in ufs
-    Assert.assertTrue((new File(mLocalUfsPath + ufsBase + "/c")).mkdirs());
-
-    // Now, 'base + /c/d' will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI(alluxioBase + "/c/d/e"));
-    checkAbsentPaths(new AlluxioURI(alluxioBase + "/c/d"));
-  }
-
-  @Test
-  public void isAbsentRemoveUfsDirectory() throws Exception {
-    String ufsBase = "/a/b";
-    String alluxioBase = "/mnt" + ufsBase;
-    // Create ufs directories
-    Assert.assertTrue((new File(mLocalUfsPath + ufsBase)).mkdirs());
-
-    // 'base + /c' will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI(alluxioBase + "/c/d/e"));
-    checkAbsentPaths(new AlluxioURI(alluxioBase + "/c"));
-
-    // delete '/a/b' from ufs
-    Assert.assertTrue((new File(mLocalUfsPath + ufsBase)).delete());
-
-    // Now, '/a/b' will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI(alluxioBase + "/c/d/e"));
-    checkAbsentPaths(new AlluxioURI(alluxioBase));
-  }
-
-  @Test
-  public void removeMountPoint() throws Exception {
-    String ufsBase = "/a/b";
-    String alluxioBase = "/mnt" + ufsBase;
-    // Create ufs directories
-    Assert.assertTrue((new File(mLocalUfsPath + ufsBase)).mkdirs());
-
-    // 'base + /c' will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI(alluxioBase + "/c/d"));
-    checkAbsentPaths(new AlluxioURI(alluxioBase + "/c"));
-
-    // Unmount
-    mUfsAbsentPathSyncCache.removeMountPoint(mMountId);
-    Assert.assertTrue(mMountTable.delete(new AlluxioURI("/mnt")));
-
-    // Re-mount the same ufs
-    long newMountId = IdUtils.getRandomNonNegativeLong();
-    mUfsManager.addMount(newMountId, mLocalUfsPath, Collections.<String, String>emptyMap());
-    mMountTable.add(new AlluxioURI("/mnt"), new AlluxioURI(mLocalUfsPath), newMountId,
-        MountOptions.defaults());
-
-    // The cache should not contain any paths now.
-    Assert.assertFalse(mUfsAbsentPathSyncCache.isAbsent(new AlluxioURI("/mnt/a/b/c/d")));
-    Assert.assertFalse(mUfsAbsentPathSyncCache.isAbsent(new AlluxioURI("/mnt/a/b/c")));
-    Assert.assertFalse(mUfsAbsentPathSyncCache.isAbsent(new AlluxioURI("/mnt/a/b")));
-    Assert.assertFalse(mUfsAbsentPathSyncCache.isAbsent(new AlluxioURI("/mnt/a")));
-    Assert.assertFalse(mUfsAbsentPathSyncCache.isAbsent(new AlluxioURI("/mnt/")));
-  }
-
-  @Test
-  public void removePath() throws Exception {
-    String ufsBase = "/a/b";
-    String alluxioBase = "/mnt" + ufsBase;
-    // Create ufs directories
-    Assert.assertTrue((new File(mLocalUfsPath + ufsBase)).mkdirs());
-
-    // 'base + /c' will be the first absent path
-    mUfsAbsentPathSyncCache.addAbsentPath(new AlluxioURI(alluxioBase + "/c/d"));
-    checkAbsentPaths(new AlluxioURI(alluxioBase + "/c"));
-
-    mUfsAbsentPathSyncCache.removeAbsentPath(new AlluxioURI(alluxioBase + "/c/d"));
-
-    Assert.assertFalse(mUfsAbsentPathSyncCache.isAbsent(new AlluxioURI("/mnt/a/b/c/d")));
-    Assert.assertFalse(mUfsAbsentPathSyncCache.isAbsent(new AlluxioURI("/mnt/a/b/c")));
-    Assert.assertFalse(mUfsAbsentPathSyncCache.isAbsent(new AlluxioURI("/mnt/a/b")));
-    Assert.assertFalse(mUfsAbsentPathSyncCache.isAbsent(new AlluxioURI("/mnt/a")));
-    Assert.assertFalse(mUfsAbsentPathSyncCache.isAbsent(new AlluxioURI("/mnt/")));
-  }
-
-  /**
-   * Checks for absent paths (descendants) and existing paths (ancestors) in the UFS.
-   *
-   * @param firstAbsent the first Alluxio path which should not exist in the UFS
-   */
-  private void checkAbsentPaths(AlluxioURI firstAbsent) throws Exception {
-    // Check for additional non-existing paths as descendants of the first absent path
-    for (int level = 1; level <= 2; level++) {
-      AlluxioURI levelUri = firstAbsent.join("level" + level);
-      for (int dir = 1; dir <= 2; dir++) {
-        AlluxioURI uri = levelUri.join("dir" + dir);
-        Assert.assertTrue(uri.toString(), mUfsAbsentPathSyncCache.isAbsent(uri));
-      }
-    }
-
-    // Check all ancestors
-    AlluxioURI existing = firstAbsent.getParent();
-    while (existing != null) {
-      Assert.assertFalse(existing.toString(), mUfsAbsentPathSyncCache.isAbsent(existing));
-      existing = existing.getParent();
-    }
-  }
-}
